a function that convert a string to an integer

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -8,31 +8,31 @@
  */
 int _atoi(char *s)
 {
-    int sign = 1;
-    int result = 0;
-    int i = 0;
+	int sign = 1;
+	long int result = 0;
+	int i = 0;
 
-    if (!s)
-        return 0;
+	if (!s)
+		return 0;
 
-    while (s[i] != '\0')
-    {
-        if (s[i] == '-')
-        {
-            sign = -sign;
-        }
-        else if (s[i] >= '0' && s[i] <= '9')
-        {
-            result = result * 10 + (s[i] - '0');
-        }
-        else if (result > 0)
-        {
-            /* stop at first non-digit after digits */
-            break;
-        }
+	while (s[i] != '\0')
+	{
+		if (s[i] == '-')
+		{
+			sign = -sign;
+		}
+		else if (s[i] >= '0' && s[i] <= '9')
+		{
+			result = result * 10 + (s[i] - '0');
+		}
+		else if (result > 0)
+		{
+			/* stop at first non-digit after digits */
+			break;
+		}
 
-        i++;
-    }
+		i++;
+	}
 
-    return (sign * result);
+	return (sign * result);
 }
